[#359] 사용하지 않는 커스텀 뷰 제거
