+++ EATSSU/App/Sources/Presentation/TabBar/CustomTabBarContainerController.swift
@@ -221,3 +221,12 @@ extension CustomTabBarContainerController: UITabBarControllerDelegate {
         return true
     }
 }
+
+extension UIImage {
+    func resized(to size: CGSize) -> UIImage {
+        let renderer = UIGraphicsImageRenderer(size: size)
+        return renderer.image { _ in
+            self.draw(in: CGRect(origin: .zero, size: size))
+        }
+    }
+}
